我要投资界面进行中

+++ app/src/main/java/com/jia16/bean/FixedEarn.java
@@ -236,14 +236,14 @@ public class FixedEarn {
     }
 
     public static class RemainingAmountBean {
-        private int amount;
+        private double amount;
         private String currency;
 
-        public int getAmount() {
+        public double getAmount() {
             return amount;
         }
 
-        public void setAmount(int amount) {
+        public void setAmount(double amount) {
             this.amount = amount;
         }
 
@@ -257,14 +257,14 @@ public class FixedEarn {
     }
 
     public static class CurrentInvestmentAmountBean {
-        private int amount;
+        private double amount;
         private String currency;
 
-        public int getAmount() {
+        public double getAmount() {
             return amount;
         }
 
-        public void setAmount(int amount) {
+        public void setAmount(double amount) {
             this.amount = amount;
         }
 
@@ -278,14 +278,14 @@ public class FixedEarn {
     }
 
     public static class AmountBean {
-        private int amount;
+        private double amount;
         private String currency;
 
-        public int getAmount() {
+        public double getAmount() {
             return amount;
         }
 
-        public void setAmount(int amount) {
+        public void setAmount(double amount) {
             this.amount = amount;
         }
 
@@ -329,14 +329,14 @@ public class FixedEarn {
         }
 
         public static class StepInvestmentAmountBean {
-            private int amount;
+            private double amount;
             private String currency;
 
-            public int getAmount() {
+            public double getAmount() {
                 return amount;
             }
 
-            public void setAmount(int amount) {
+            public void setAmount(double amount) {
                 this.amount = amount;
             }
 
@@ -350,14 +350,14 @@ public class FixedEarn {
         }
 
         public static class MinimumInvestmentAmountBean {
-            private int amount;
+            private double amount;
             private String currency;
 
-            public int getAmount() {
+            public double getAmount() {
                 return amount;
             }
 
-            public void setAmount(int amount) {
+            public void setAmount(double amount) {
                 this.amount = amount;
             }
