+++ app/src/main/java/com/jia16/bean/InvestConstant.java
@@ -206,14 +206,14 @@ public class InvestConstant {
     }
 
     public static class AmountBean {
-        private int amount;
+        private double amount;
         private String currency;
 
-        public int getAmount() {
+        public double getAmount() {
             return amount;
         }
 
-        public void setAmount(int amount) {
+        public void setAmount(double amount) {
             this.amount = amount;
         }
 
@@ -235,14 +235,14 @@ public class InvestConstant {
     }
 
     public static class RemainingAmountBean {
-        private int amount;
+        private double amount;
         private String currency;
 
-        public int getAmount() {
+        public double getAmount() {
             return amount;
         }
 
-        public void setAmount(int amount) {
+        public void setAmount(double amount) {
             this.amount = amount;
         }
 
@@ -264,14 +264,14 @@ public class InvestConstant {
     }
 
     public static class CurrentInvestmentAmountBean {
-        private int amount;
+        private double amount;
         private String currency;
 
-        public int getAmount() {
+        public double getAmount() {
             return amount;
         }
 
-        public void setAmount(int amount) {
+        public void setAmount(double amount) {
             this.amount = amount;
         }
 
@@ -337,14 +337,14 @@ public class InvestConstant {
         }
 
         public static class MinimumInvestmentAmountBean {
-            private int amount;
+            private double amount;
             private String currency;
 
-            public int getAmount() {
+            public double getAmount() {
                 return amount;
             }
 
-            public void setAmount(int amount) {
+            public void setAmount(double amount) {
                 this.amount = amount;
             }
 
@@ -358,14 +358,14 @@ public class InvestConstant {
         }
 
         public static class StepInvestmentAmountBean {
-            private int amount;
+            private double amount;
             private String currency;
 
-            public int getAmount() {
+            public double getAmount() {
                 return amount;
             }
 
-            public void setAmount(int amount) {
+            public void setAmount(double amount) {
                 this.amount = amount;
             }
 
@@ -379,14 +379,14 @@ public class InvestConstant {
         }
 
         public static class MaximumInvestmentAmountBean {
-            private int amount;
+            private double amount;
             private String currency;
 
-            public int getAmount() {
+            public double getAmount() {
                 return amount;
             }
 
-            public void setAmount(int amount) {
+            public void setAmount(double amount) {
                 this.amount = amount;
             }
